Grouping combine_datasets function with functions in analyse._alignment.py
